Merged PR 4970: IG Cloud Deferred Revenue
This change, parameterises between local dev execution and running in the proper env, and adds in a cell for the derivation of DetID's (thats Detail Record IDs) that have deferred revenue.

The deferred dates are derived and stored along with a PeriodCount. The Revenue for these DetIDs is simply divided by the PeriodCount and records created for the future periods, again using Revenue / PeriodCount.

Costs and GP1 are also calculated the same way as these needed to be deferred as well.

Related work items: #20693