Merged PR 5085: Added dummy account vendor override codes to fix rec issues
TLDR: Added dummy codes to the account lookup table to be used to override vendor codes when accounts should be aggregated together before any aggregation is done at the vendor level. This is to help reconcile the data we're outputting to the reporting layer against what is expected of Tagetik Data.

Background:

For the Tagetik outputs; revenue from some accounts which count as services are isolated and aggregated without vendor consideration. Some of the discrepancies we were seeing was due to these services being aggregated within the vendor when it shouldn't be.

To correct this in a simple a way as possible I've added in dummy vendor codes to the account lookup table in the gold layer, then going into the platinum layer the vendor codes for these accounts are overriden. That way they can be aggregated and presented as if these services are another vendor. Thought this was the easiest way without having to make changes within the PowerBI layer. I've added the codes to the vendor mapping spreadsheet as well so the proper names are presented in PowerBI.

Related work items: #21453